test: added tests for the isInteger type
